requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.1.1
RapidFuzz==3.14.0
rsa==4.9.1
Telethon==1.41.0
typing_extensions==4.15.0
yarl==1.20.1
